Refactor out into recompute_head.rs
