+++ beacon_node/beacon_chain/src/builder.rs
@@ -718,9 +718,17 @@ where
         let fork_choice_view = fork_choice.cached_fork_choice_view();
         let canonical_head = CanonicalHead {
             fork_choice,
-            head_proto_block,
             fork_choice_view,
-            head_snapshot: head_snapshot.clone(),
+            head_proposer_shuffling_decision_root: head_snapshot
+                .beacon_state
+                .proposer_shuffling_decision_root(head_snapshot.beacon_block_root)
+                .map_err(|e| format!("Failed to determine shuffling root: {:?}", e))?,
+            head_random: *head_snapshot
+                .beacon_state
+                .get_randao_mix(head_snapshot.beacon_state.current_epoch())
+                .map_err(|e| format!("Failed to determine randao mix: {:?}", e))?,
+            head_execution_status: head_proto_block.execution_status,
+            head_snapshot,
         };
 
         let beacon_chain = BeaconChain {
