+++ beacon_node/beacon_chain/src/lib.rs
@@ -29,6 +29,7 @@ mod persisted_beacon_chain;
 mod persisted_fork_choice;
 mod pre_finalization_cache;
 pub mod proposer_prep_service;
+mod recompute_head;
 pub mod schema_change;
 mod shuffling_cache;
 mod snapshot_cache;
